Update Gradle, Kotlin, and Gradle plug-in for Android
Trying to build/install this application in a new installation of Android Studio yielded:

> Unable to determine application id: com.android.tools.idea.run.ApkProvisionException: Error loading build artifacts from:<snip>\sms-matrix\app\build\outputs\apk\debug\output-metadata.json

This updates the versions of applications central to building the Android project to make it compatible. This does not (intentionally, at least) have any functional impact on the application at all.

I opted not to jump to Gradle 7.0 as it has just recently been released and increasing 1 major version is change enough for one PR.